+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     - Renamed Dataset types method to columnTypes
     - Efficient serialization of Word Count Vectorizer
     - Decoupled Persistable interface from Learner
+    - Moved Gower Distance kernel to Extras package
 
 - 0.1.0-rc2
     - Persistent Model now implements Verbose interface